backend/requirements.txt
flake8==7.2.0
coverage==7.9.1
requests==2.32.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis==6.2.0
channels==4.2.2
channels-redis==4.2.1
daphne==4.2.1
celery==5.5.3
flower==2.0.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-filter==25.1
django-cors-headers==4.7.0
django-redis==6.0.0
django-celery-results==2.6.0
django-import-export==4.3.8
